services/finnhub_data/pyproject.toml
[tool.poetry]
name = "finnhub_data"
version = "0.1.8"
description = ""
authors = ["ja185332 <ja185332.ncr.com>"]

[tool.poetry.dependencies]
python = "^3.8"
finnhub-python = "^2.4.13"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
